+++ azurerm/internal/services/network/virtual_network_gateway_data_source.go
@@ -48,6 +48,11 @@ func dataSourceArmVirtualNetworkGateway() *schema.Resource {
 				Computed: true,
 			},
 
+			"enable_private_ip_address": {
+				Type:     schema.TypeBool,
+				Computed: true,
+			},
+
 			"active_active": {
 				Type:     schema.TypeBool,
 				Computed: true,
@@ -199,6 +204,25 @@ func dataSourceArmVirtualNetworkGateway() *schema.Resource {
 				},
 			},
 
+			"custom_route": {
+				Type:     schema.TypeList,
+				Optional: true,
+				ForceNew: true,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"address_prefixes": {
+							Type:     schema.TypeSet,
+							Optional: true,
+							ForceNew: true,
+							Elem: &schema.Schema{
+								Type: schema.TypeString,
+							},
+						},
+					},
+				},
+			},
+
 			"default_local_network_gateway_id": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -239,6 +263,7 @@ func dataSourceArmVirtualNetworkGatewayRead(d *schema.ResourceData, meta interfa
 
 		d.Set("type", string(gw.GatewayType))
 		d.Set("enable_bgp", gw.EnableBgp)
+		d.Set("enable_private_ip_address", gw.EnablePrivateIPAddress)
 		d.Set("active_active", gw.ActiveActive)
 		d.Set("generation", string(gw.VpnGatewayGeneration))
 
@@ -266,6 +291,10 @@ func dataSourceArmVirtualNetworkGatewayRead(d *schema.ResourceData, meta interfa
 		bgpSettingsFlat := flattenArmVirtualNetworkGatewayDataSourceBgpSettings(gw.BgpSettings)
 		if err := d.Set("bgp_settings", bgpSettingsFlat); err != nil {
 			return fmt.Errorf("Error setting `bgp_settings`: %+v", err)
+		}
+
+		if err := d.Set("custom_route", flattenArmVirtualNetworkGatewayAddressSpace(gw.CustomRoutes)); err != nil {
+			return fmt.Errorf("setting `custom_route`: %+v", err)
 		}
 	}