+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 2.35.0 (Unreleased)
 
+FEATURES:
+
+* **New Resource:** `azurerm_firewall_policy_rule_collection_group` [GH-8603]
+
 ## 2.34.0 (October 29, 2020)
 
 UPGRADE NOTES